# fix autotest bug and optimize eager test speed in ci