--- a/python/oneflow/test/modules/test_amax.py
+++ b/python/oneflow/test/modules/test_amax.py
@@ -104,7 +104,7 @@ class TestAmax(flow.unittest.TestCase):
         for arg in GenArgList(arg_dict):
             arg[0](test_case, *arg[1:])
 
-    @autotest()
+    @autotest(n=5)
     def test_amax_with_random_data_single_dim(test_case):
         device = random_device()
         ndim = random(1, 6).to(int)
@@ -112,7 +112,7 @@ class TestAmax(flow.unittest.TestCase):
         y = torch.amax(x, dim=random(0, ndim), keepdim=random().to(bool))
         return y
 
-    @autotest()
+    @autotest(n=5)
     def test_amax_with_random_data_empty_dim(test_case):
         device = random_device()
         ndim = random(1, 6).to(int)
@@ -120,7 +120,7 @@ class TestAmax(flow.unittest.TestCase):
         y = torch.amax(x, dim=None, keepdim=random().to(bool))
         return y
 
-    @autotest()
+    @autotest(n=5)
     def test_amax_with_random_data_multi_dims(test_case):
         device = random_device()
         ndim = random(2, 6).to(int)