--- a/python/oneflow/test/modules/test_ceil.py
+++ b/python/oneflow/test/modules/test_ceil.py
@@ -25,28 +25,28 @@ from oneflow.test_utils.automated_test_util import *
 
 @flow.unittest.skip_unless_1n1d()
 class TestCeilModule(flow.unittest.TestCase):
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_ceil_flow_with_random_data(test_case):
         device = random_device()
         input = random_tensor().to(device)
         y = torch.ceil(input)
         return y
 
-    @autotest(check_graph=True)
+    @autotest(n=5)
     def test_ceil_flow_with_random_0d_data(test_case):
         device = random_device()
         input = random_tensor(ndim=0).to(device)
         y = torch.ceil(input)
         return y
 
-    @autotest(auto_backward=False, check_graph=True)
+    @autotest(n=5, auto_backward=False, check_graph=True)
     def test_ceil_with_0_size_data(test_case):
         device = random_device()
         x = random_tensor(4, 2, 1, 0, 3).to(device)
         y = torch.ceil(x)
         return y
 
-    @autotest(auto_backward=False, check_graph=True)
+    @autotest(n=5, auto_backward=False, check_graph=True)
     def test_ceil_with_0shape_0d_data(test_case):
         device = random_device()
         x = random_tensor(ndim=0).to(device)