--- a/python/oneflow/test/modules/test_eq.py
+++ b/python/oneflow/test/modules/test_eq.py
@@ -28,7 +28,7 @@ from oneflow.test_utils.automated_test_util import *
 
 @flow.unittest.skip_unless_1n1d()
 class TestEq(flow.unittest.TestCase):
-    @autotest(auto_backward=False, check_graph=True)
+    @autotest(n=5, auto_backward=False, check_graph=True)
     def test_eq_with_0_size_data(test_case):
         device = random_device()
         x = random_tensor(3, 2, 0, 3).to(device)
@@ -36,7 +36,7 @@ class TestEq(flow.unittest.TestCase):
         z = torch.eq(x, y)
         return z
 
-    @autotest(auto_backward=False, check_graph=True)
+    @autotest(n=5, auto_backward=False, check_graph=True)
     def test_eq_with_0shape_0d_data(test_case):
         device = random_device()
         x = random_tensor(ndim=0).to(device)
@@ -44,16 +44,15 @@ class TestEq(flow.unittest.TestCase):
         z = torch.eq(x, y)
         return z
 
-    @autotest(auto_backward=False, check_graph=True)
+    @autotest(n=5, auto_backward=False, check_graph=True)
     def test_flow_eq_with_random_data(test_case):
         device = random_device()
         shape = random_tensor().oneflow.shape
-        print(*shape)
         x = random_tensor(len(shape), *shape, requires_grad=False).to(device)
         y = random_tensor(len(shape), *shape, requires_grad=False).to(device)
         return torch.eq(x, y)
 
-    @autotest(auto_backward=False, check_graph=True)
+    @autotest(n=5, auto_backward=False, check_graph=True)
     def test_flow_eq_with_random_0d_data(test_case):
         device = random_device()
         shape = random_tensor().oneflow.shape
@@ -61,14 +60,14 @@ class TestEq(flow.unittest.TestCase):
         y = random_tensor(ndim=0, requires_grad=False).to(device)
         return torch.eq(x, y)
 
-    @autotest(auto_backward=False, check_graph=True)
+    @autotest(n=5, auto_backward=False, check_graph=True)
     def test_flow_eq_with_same_random_data(test_case):
         device = random_device()
         shape = random_tensor().oneflow.shape
         x = random_tensor(len(shape), *shape, requires_grad=False).to(device)
         return torch.eq(x, x)
 
-    @autotest(auto_backward=False, check_graph=True)
+    @autotest(n=5, auto_backward=False, check_graph=True)
     def test_flow_eq_bool_with_random_data(test_case):
         device = random_device()
         shape = random_tensor().oneflow.shape
@@ -80,7 +79,7 @@ class TestEq(flow.unittest.TestCase):
         )
         return torch.eq(x, y)
 
-    @autotest(auto_backward=False, check_graph=True)
+    @autotest(n=5, auto_backward=False, check_graph=True)
     def test_flow_eq_with_same_random_0d_data(test_case):
         device = random_device()
         shape = random_tensor().oneflow.shape