- [x] 修复AutoTest打印fake_program时一个判断条件写错的bug。
- [x] 减少 test_adaptive.py 运行时间。25s->11s。
- [x] 减少test_addmm.py 运行时间。12s->8s。
- [x] 减少test_amax.py 运行时间。17s->8s.
- [x] 减少test_amin.py 运行时间。16s->8s。
- [x] 减少 test_avgpool.py 运行时间。28s->11s。
- [x] 减少 test_ceil.py 运行时间。20s->9s。
- [x] 减少 test_clamp.py 运行时间。30s->11s。
- [x] 减少test_concat.py 运行时间。20s->12s。
- [x] 减少test_conv1d.py 运行时间。26s->10s。
- [x] 减少test_conv2d.py 运行时间。23s->11s。
- [x] 减少test_div.py 运行时间。20s->9s。
- [x]  减少 test_eq.py 运行时间。32s->12s。
- [x] 减少 test_erf.py / test_erfc.py 运行时间。13s->7s。
- [x] 减少test_expm1.py 运行时间。16s->8s。
- [x] 减少 test_flatten.py 运行时间。20s->9s。
- [x] 减少test_flip.py 运行时间。16s->11s。
- [x] 减少 test_greater.py 运行时间。26s->12s。
- [x] 减少 test_linear.py 运行时间。21s->9s。
- [x] 减少 test_linspace.py 运行时间。16s->7s。
- [x] 减少 test_norm.py 运行时间。25s->11s。
- [x] 减少 test_prod.py 运行时间。20s->9s。
- [x] 减少 test_reshape.py 运行时间。20s->9s。
- [x] 减少test_sign.py 运行时间。15s->9s。
- [x] 减少test_split.py 运行时间。40s->14s。
- [x] 减少test_tril.py 运行时间。20s->9s。
- [x] 减少test_where.py 运行时间。137s->37s。

以上CI里耗时大于10s的Eager Test可以优化到原来总执行时间的 (25+12+17+16+28+20+30+20+26+23+20+32+13+16+20+16+26+21+16+25+20+20+15+40+20+137) /(11+8+8+8+11+9+11+12+10+11+9+12+7+8+9+11+12+9+7+11+9+9+9+14+9+37)  = 674/291 等于2.3倍。